Componetize calculator results section
