+++ src/components/CalculationResults/__test__/CalculationResults.test.tsx
@@ -1,0 +1,12 @@
+/**
+ * @file CalculationResults.test.tsx
+ * @description
+ * @author Dushan Ranasinghage
+ * @copyright Copyright 2026 - All Rights Reserved.
+ */
+
+describe("CalculationResults Component", () => {
+  test("placeholder test", () => {
+    expect(true).toBe(true);
+  });
+});
